+++ src/betterdiscord/builtins/developer/recovery.tsx
@@ -9,7 +9,7 @@ import Toasts from "@stores/toasts";
 import pluginmanager from "@modules/pluginmanager";
 import IPC from "@modules/ipc";
 import Modals from "@ui/modals";
-import {getByKeys, getByPrototypes, getByStrings, getMangled} from "@webpack";
+import {getByPrototypes, getByStrings} from "@webpack";
 import NotificationUIInstance from "@ui/notifications";
 import config from "@stores/config";
 import {Logo} from "@ui/logo";
@@ -20,8 +20,7 @@ const TEST_PLUGIN_REGEX = /betterdiscord:\/\/(plugins)\/(.*?).(\w+).js/;
 
 // TODO: arven if you get a chance
 async function attemptRecovery() {
-    const transitionTo = getByStrings(["transitionTo - Transitioning to"], {searchExports: true});
-    const modalModule = getMangled(`,["contextKey"]),`, {CloseAllModals: x => x.toString?.()?.includes(".key,") && x.toString?.()?.includes("getState();")});
+    const transitionTo = getByStrings(["transitionTo - Transitioning to"], {searchExports: true, firstId: 976860, cacheId: "core-recovery-transitionTo"});
 
     const recoverySteps = [
         {
@@ -37,7 +36,7 @@ async function attemptRecovery() {
             errorMessage: "Failed to route to main channel"
         },
         {
-            action: () => modalModule?.CloseAllModals(),
+            action: () => Modals.ModalActions?.closeAllModals?.(),
             errorMessage: "Failed to close all modals"
         }
     ];
@@ -176,7 +175,6 @@ export default new class Recovery extends Builtin {
 
     async enabled() {
         this.patchErrorBoundry();
-        this.parseModule = getByKeys(["defaultRules", "parse"]);
     }
 
     async disabled() {
@@ -200,7 +198,7 @@ export default new class Recovery extends Builtin {
     }
 
     patchErrorBoundry() {
-        const mod = getByPrototypes(["_handleSubmitReport"]);
+        const mod = getByPrototypes(["_handleSubmitReport"], {firstId: 670735, cacheId: "core-recovery-ErrorBoundary"});
 
         this.after(mod?.prototype, "render", (instance, args, retValue) => {
             if (!Settings.get(this.collection, this.category, this.id)) return;
@@ -209,7 +207,7 @@ export default new class Recovery extends Builtin {
             if (!buttons) return;
 
             const errorStack = instance.state;
-            const parsedError = errorStack ? this.parseModule.parse(`\`\`\`${errorStack.error?.stack}\n\n${errorStack.info?.componentStack}\`\`\``) : null;
+            const parsedError = errorStack ? DiscordModules.SimpleMarkdownWrapper.parse(`\`\`\`${errorStack.error?.stack}\n\n${errorStack.info?.componentStack}\`\`\``) : null;
 
             const foundIssue = TEST_PLUGIN_REGEX.exec(errorStack.error?.stack);
             let pluginInfo = null;